myproject/requirements.txt
django
djangorestframework
opencv-python